requirements.txt
flask
oracledb
transformers
torch
python-dotenv
pandas
